pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "Clisonix-cloud"
version = "0.1.0"
description = "Cloud platform for Clisonix EEG-to-Audio and Brain Symphony"
authors = [{name="Ledjan Ahmati"}]
license = {text = "MIT"}
requires-python = ">=3.9"

dependencies = [
    # FastAPI and server
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    
    # Database
    "sqlalchemy[asyncio]>=2.0.23",
    "asyncpg>=0.29.0",
    "alembic>=1.12.1",
    
    # Redis
    "redis[hiredis]>=5.0.1",
    
    # Authentication & Security
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    
    # Stripe payments
    "stripe>=7.8.0",
    
    # Settings management
    "pydantic[email]>=2.5.0",
    "pydantic-settings>=2.1.0",
    
    # File storage
    "boto3>=1.34.0",
    "botocore>=1.34.0",
    
    # Scientific computing
    "numpy>=1.25.0",
    "scipy>=1.11.0",
    "librosa>=0.10.1",
    "mne>=1.5.0",
    
    # System monitoring
    "psutil>=5.9.0",
    
    # Monitoring & Observability
    "opentelemetry-api>=1.21.0",
    "opentelemetry-sdk>=1.21.0",
    "opentelemetry-exporter-jaeger>=1.21.0",
    "opentelemetry-instrumentation-fastapi>=0.42b0",
    "opentelemetry-instrumentation-redis>=0.42b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.42b0",
    
    # Task queue
    "celery[redis]>=5.3.4",
    
    # Email
    "fastapi-mail>=1.4.1",
    
    # Validation & Utilities
    "email-validator>=2.1.0",
    "python-dateutil>=2.8.2",
    "pytz>=2023.3",
    
    # Logging & Monitoring
    "structlog>=23.2.0",
    "sentry-sdk[fastapi]>=1.38.0",
    
    # Development
    "python-dotenv>=1.0.0"
]
